+++ src/onegov.form/onegov/form/__init__.py
@@ -2,7 +2,11 @@ import logging
 log = logging.getLogger('onegov.form')  # noqa
 log.addHandler(logging.NullHandler())  # noqa
 
-from onegov.form.collection import FormCollection
+from onegov.form.collection import (
+    FormCollection,
+    FormSubmissionCollection,
+    FormDefinitionCollection
+)
 from onegov.form.core import Form, with_options
 from onegov.form.display import render_field
 from onegov.form.models import (
@@ -15,6 +19,8 @@ from onegov.form.models import (
 __all__ = [
     'Form',
     'FormCollection',
+    'FormDefinitionCollection',
+    'FormSubmissionCollection',
     'FormDefinition',
     'FormSubmission',
     'PendingFormSubmission',